+++ src/main/kotlin/lotto/domain/generator/RandomLottoGenerator.kt
@@ -1,7 +1,7 @@
-package lotto.generator
+package lotto.domain.generator
 
-import lotto.Lotto
-import lotto.LottoNumber
+import lotto.domain.Lotto
+import lotto.domain.LottoNumber
 
 class RandomLottoGenerator : LottoGenerator {
     override fun get(): Lotto {
