+++ src/test/kotlin/lotto/domain/LottoMachineTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.matchers.shouldBe
 import lotto.fixture.generatorWithParameter
