+++ src/test/kotlin/lotto/domain/MatchTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
