+++ src/test/kotlin/lotto/fixture/LottoGeneratorFixture.kt
@@ -1,7 +1,7 @@
 package lotto.fixture
 
-import lotto.Lotto
-import lotto.generator.LottoGenerator
+import lotto.domain.Lotto
+import lotto.domain.generator.LottoGenerator
 
 fun generatorWithParameter(numbers: List<Int>): LottoGenerator {
     return LottoGenerator { Lotto.from(numbers) }
